refactor(operator): colocate constructor options
Keep each constructor option declaration next to its owning implementation and remove the shared options file.
