+++ common/go/operator/operator.go
@@ -8,6 +8,8 @@ import (
 	"go.uber.org/zap"
 	"golang.org/x/sync/errgroup"
 	"google.golang.org/grpc"
+
+	"github.com/yanet-platform/yanet2/common/go/xcfg"
 )
 
 // Runner is a long-running goroutine driven by the operator's errgroup.
@@ -27,6 +29,124 @@ type PreRun func(ctx context.Context) error
 //
 // The name is used for gateway registration heartbeats.
 type ServiceRegistrar func(server *grpc.Server) string
+
+type grpcServerOption struct {
+	// Config controls where the embedded server listens and advertises itself.
+	Config GRPCServerConfig
+	// Services are registered with the embedded server at construction.
+	Services []ServiceRegistrar
+}
+
+type options struct {
+	// Reconcile supplies timing defaults consumed by the reconcile loop.
+	Reconcile ReconcileConfig
+	// Register supplies the heartbeat interval for gateway registration.
+	Register RegisterConfig
+	// Gateways receive registration heartbeats when the server is enabled.
+	Gateways []GatewayConfig
+	// Workers run alongside the reconciler in the shared error group.
+	Workers []Runner
+	// PreRun initializes state before concurrent work starts.
+	PreRun PreRun
+	// Metrics receives reconcile-loop lifecycle events.
+	Metrics ReconcilerMetricsObserver
+	// GRPCServer is nil when the embedded server is disabled.
+	GRPCServer *grpcServerOption
+	// Log is the fallback logger for operator components.
+	Log *zap.Logger
+}
+
+func newOptions() *options {
+	return &options{
+		Reconcile: ReconcileConfig{
+			Interval:       xcfg.MustNonZero(DefaultReconcileInterval),
+			InitialBackoff: xcfg.MustNonZero(DefaultReconcileInitialBackoff),
+			MaxBackoff:     xcfg.MustNonZero(DefaultReconcileMaxBackoff),
+		},
+		PreRun:  func(context.Context) error { return nil },
+		Metrics: NoopReconcilerMetricsObserver{},
+		Log:     zap.NewNop(),
+	}
+}
+
+// Option configures the Operator.
+type Option func(*options)
+
+// WithLog sets the logger used by the Operator and its sub-components
+// when no more specific logger is supplied.
+func WithLog(log *zap.Logger) Option {
+	return func(o *options) {
+		o.Log = log
+	}
+}
+
+// WithReconcile sets the reconcile-loop timing parameters.
+//
+// Zero-valued fields set the corresponding default value.
+func WithReconcile(cfg ReconcileConfig) Option {
+	return func(o *options) {
+		if cfg.Interval.Unwrap() > 0 {
+			o.Reconcile.Interval = cfg.Interval
+		}
+		if cfg.InitialBackoff.Unwrap() > 0 {
+			o.Reconcile.InitialBackoff = cfg.InitialBackoff
+		}
+		if cfg.MaxBackoff.Unwrap() > 0 {
+			o.Reconcile.MaxBackoff = cfg.MaxBackoff
+		}
+	}
+}
+
+// WithGateways enables gateway registration against the supplied gateways.
+//
+// An empty gateway list leaves registration disabled. Registration also
+// requires an embedded gRPC server.
+func WithGateways(register RegisterConfig, gateways ...GatewayConfig) Option {
+	return func(o *options) {
+		o.Register = register
+		o.Gateways = gateways
+	}
+}
+
+// WithGRPCServer enables an embedded gRPC server on the configured endpoint.
+//
+// The server exposes the supplied services independently of gateway
+// registration, so operators may serve metrics or APIs without registering.
+// When this option is not supplied, the operator does not bind a listener.
+func WithGRPCServer(cfg GRPCServerConfig, services ...ServiceRegistrar) Option {
+	return func(o *options) {
+		o.GRPCServer = &grpcServerOption{
+			Config:   cfg,
+			Services: services,
+		}
+	}
+}
+
+// WithWorkers attaches additional long-running goroutines to the
+// operator's errgroup.
+//
+// Returning a non-nil error from any worker tears down the operator.
+func WithWorkers(workers ...Runner) Option {
+	return func(o *options) {
+		o.Workers = workers
+	}
+}
+
+// WithPreRun registers a hook executed once before any goroutine starts.
+//
+// Useful for seeding the source from static configuration.
+func WithPreRun(fn PreRun) Option {
+	return func(o *options) {
+		o.PreRun = fn
+	}
+}
+
+// WithMetrics attaches the metrics observer for the reconcile loop.
+func WithMetrics(observer ReconcilerMetricsObserver) Option {
+	return func(o *options) {
+		o.Metrics = observer
+	}
+}
 
 // Operator is the generic operator skeleton.
 //
